docs: Add missing XML docs
