+++ src/Uno.Extensions.Reactive/Core/Feed.T.cs
@@ -10,6 +10,7 @@ namespace Uno.Extensions.Reactive;
 /// <summary>
 /// Provides a set of static methods to create and manipulate <see cref="IFeed{T}"/>.
 /// </summary>
+/// <typeparam name="T">The type of the data.</typeparam>
 public static class Feed<T> // We set the T on the class to it greatly helps type inference of factory delegates
 {
 	/// <summary>
