+++ src/Uno.Extensions.Reactive/Core/ListFeed.Extensions.cs
@@ -2,15 +2,14 @@
 using System.Collections.Generic;
 using System.Collections.Immutable;
 using System.Linq;
-using System.Runtime.CompilerServices;
 using System.Threading;
-using System.Threading.Tasks;
-using Uno.Extensions.Collections.Tracking;
-using Uno.Extensions.Reactive.Core;
 using Uno.Extensions.Reactive.Utils;
 
 namespace Uno.Extensions.Reactive;
 
+/// <summary>
+/// Provides a set of static methods to create and manipulate <see cref="IListFeed{T}"/>.
+/// </summary>
 public static partial class ListFeed
 {
 	#region Sources
@@ -47,21 +46,37 @@ public static partial class ListFeed
 	#endregion
 
 	#region Convertions
+	/// <summary>
+	/// Wraps a feed of list into a <see cref="IListFeed{T}"/>.
+	/// </summary>
+	/// <typeparam name="TItem">Type of items in the list.</typeparam>
+	/// <param name="source">The source list stream to wrap.</param>
+	/// <returns>A <see cref="IListFeed{T}"/> that wraps the given source data stream of list.</returns>
 	public static IListFeed<TItem> AsListFeed<TItem>(this IFeed<IImmutableList<TItem>> source)
 	{
 		// Note: We are not attaching the "ListFeed" as we always un-wrap them and we attach other operator on the underlying Source.
-		
-
-
 		return new ListFeed<TItem>(source);
 	}
 
+	/// <summary>
+	/// Wraps a feed of list into a <see cref="IListFeed{T}"/>.
+	/// </summary>
+	/// <typeparam name="TCollection">Type of the items collection.</typeparam>
+	/// <typeparam name="TItem">Type of items in the list.</typeparam>
+	/// <param name="source">The source list stream to wrap.</param>
+	/// <returns>A <see cref="IListFeed{T}"/> that wraps the given source data stream.</returns>
 	public static IListFeed<TItem> AsListFeed<TCollection, TItem>(
 		this IFeed<TCollection> source)
 		where TCollection : IImmutableList<TItem>
 		// Note: We are not attaching the "ListFeed" as we always un-wrap them and we attach other operator on the underlying Source.
 		=> new ListFeed<TItem>(source.Select(list => list.ToImmutableList() as IImmutableList<TItem>));
 
+	/// <summary>
+	/// Unwraps a <see cref="IListFeed{T}"/> to get the source feed of list.
+	/// </summary>
+	/// <typeparam name="TItem">Type of items in the list.</typeparam>
+	/// <param name="source">The source list stream to wrap.</param>
+	/// <returns>The source data stream of list of the given <see cref="IListFeed{T}"/>.</returns>
 	public static IFeed<IImmutableList<TItem>> AsFeed<TItem>(
 		this IListFeed<TItem> source)
 		=> source is IListFeedWrapper<TItem> wrapper
@@ -70,10 +85,10 @@ public static partial class ListFeed
 			// All subsequent operators are going to be attached to the newly created wrap Feed (i.e. not ListFeed)
 			: AttachedProperty.GetOrCreate(source, typeof(TItem), WrapListFeed);
 
-	public static IListState<TItem> AsListState<TItem>(
-		this IState<IImmutableList<TItem>> source)
-		// Note: We are not attaching the "ListFeed" as we always un-wrap them and we attach other operator on the underlying Source.
-		=> new ListState<TItem>(source);
+	//public static IListState<TItem> AsListState<TItem>(
+	//	this IState<IImmutableList<TItem>> source)
+	//	// Note: We are not attaching the "ListFeed" as we always un-wrap them and we attach other operator on the underlying Source.
+	//	=> new ListState<TItem>(source);
 
 	//public static IState<IImmutableList<TItem>> AsState<TItem>(
 	//	this IListState<TItem> source)
@@ -91,15 +106,15 @@ public static partial class ListFeed
 	#endregion
 
 	#region Operators
-	public static IListFeed<TSource> Where<TSource>(
-		this IListFeed<TSource> source,
-		Predicate<TSource?> predicate)
-		=> default!;
+	//public static IListFeed<TSource> Where<TSource>(
+	//	this IListFeed<TSource> source,
+	//	Predicate<TSource?> predicate)
+	//	=> default!;
 
-	public static IListFeed<TResult> Select<TSource, TResult>(
-		this IListFeed<TSource> source,
-		Func<TSource?, TResult?> selector)
-		=> default!;
+	//public static IListFeed<TResult> Select<TSource, TResult>(
+	//	this IListFeed<TSource> source,
+	//	Func<TSource?, TResult?> selector)
+	//	=> default!;
 
 	/*
 	public static IFeed<TResult> SelectAsync<TSource, TResult>(
@@ -109,33 +124,3 @@ public static partial class ListFeed
 	*/
 	#endregion
 }
-
-//internal class CollectionTrackerFeed<T> : IFeed<T>
-//{
-//	private readonly IFeed<IImmutableList<T>> _source;
-
-//	public CollectionTrackerFeed(IFeed<IImmutableList<T>> source)
-//	{
-//		_source = source;
-//	}
-
-//	/// <inheritdoc />
-//	public async IAsyncEnumerable<Message<T>> GetSource(SourceContext context, [EnumeratorCancellation] CancellationToken ct = default)
-//	{
-//		var currentItems = default(IImmutableList<T>);
-//		var tracker = new CollectionTracker();
-//		await foreach (var message in _source.GetSource(context, ct).WithCancellation(ct).ConfigureAwait(false))
-//		{
-//			if (message.Changes.Contains(MessageAxis.Data))
-//			{
-//				if (message.Changes.)
-
-//				var changes = tracker.GetChanges()
-//			}
-
-//		}
-
-
-//		throw new NotImplementedException();
-//	}
-//}
